+++ magpie/fitness/__init__.py
@@ -1,6 +1,6 @@
 from .bloat import BloatCharsFitness, BloatLinesFitness, BloatWordsFitness
 from .gnu_time import GnuTimeTemplatedFitness
-from .output import OutputFitness
+from .output import OutputFitness, OutputTemplatedFitness
 from .perf import PerfTemplatedFitness
 from .repair import RepairFitness
 from .time import (